Refactor Babybear
Moving all BabyBear specifics into the BabyBear crate.

This will allow us to access the internals without having to call to_monty or from_monty.
