+++ baby-bear/src/lib.rs
@@ -4,8 +4,10 @@ extern crate alloc;
 
 mod baby_bear;
 mod extension;
+mod poseidon2;
 
 pub use baby_bear::*;
+pub use poseidon2::*;
 
 #[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
 mod aarch64_neon;
