+++ poseidon2/src/lib.rs
@@ -8,20 +8,18 @@
 
 extern crate alloc;
 
-mod babybear;
 mod diffusion;
 mod goldilocks;
 mod matrix;
 use alloc::vec::Vec;
 
-pub use babybear::DiffusionMatrixBabybear;
 pub use diffusion::DiffusionPermutation;
 pub use goldilocks::DiffusionMatrixGoldilocks;
+
 use matrix::Poseidon2MEMatrix;
 use p3_field::{AbstractField, PrimeField};
 use p3_symmetric::{CryptographicPermutation, Permutation};
-use rand::distributions::Standard;
-use rand::prelude::Distribution;
+use rand::distributions::{Distribution, Standard};
 use rand::Rng;
 
 const SUPPORTED_WIDTHS: [usize; 8] = [2, 3, 4, 8, 12, 16, 20, 24];
@@ -167,21 +165,18 @@ mod tests {
     use alloc::vec::Vec;
 
     use ark_ff::{BigInteger, PrimeField};
-    use p3_baby_bear::BabyBear;
     use p3_field::AbstractField;
     use p3_goldilocks::Goldilocks;
     use p3_symmetric::Permutation;
     use rand::Rng;
-    use zkhash::fields::babybear::FpBabyBear;
     use zkhash::fields::goldilocks::FpGoldiLocks;
-    use zkhash::poseidon2::poseidon2::Poseidon2 as Poseidon2Ref;
-    use zkhash::poseidon2::poseidon2_instance_babybear::{POSEIDON2_BABYBEAR_16_PARAMS, RC16};
+    use zkhash::poseidon2::poseidon2::Poseidon2 as Poseidon2Ref;  
     use zkhash::poseidon2::poseidon2_instance_goldilocks::{
         POSEIDON2_GOLDILOCKS_12_PARAMS, POSEIDON2_GOLDILOCKS_8_PARAMS, RC12, RC8,
     };
 
     use crate::goldilocks::DiffusionMatrixGoldilocks;
-    use crate::{DiffusionMatrixBabybear, Poseidon2};
+    use crate::{Poseidon2};
 
     fn goldilocks_from_ark_ff(input: FpGoldiLocks) -> Goldilocks {
         let as_bigint = input.into_bigint();
@@ -189,14 +184,6 @@ mod tests {
         as_bytes.resize(8, 0);
         let as_u64 = u64::from_le_bytes(as_bytes[0..8].try_into().unwrap());
         Goldilocks::from_wrapped_u64(as_u64)
-    }
-
-    fn babybear_from_ark_ff(input: FpBabyBear) -> BabyBear {
-        let as_bigint = input.into_bigint();
-        let mut as_bytes = as_bigint.to_bytes_le();
-        as_bytes.resize(4, 0);
-        let as_u32 = u32::from_le_bytes(as_bytes[0..4].try_into().unwrap());
-        BabyBear::from_wrapped_u32(as_u32)
     }
 
     #[test]
@@ -322,69 +309,6 @@ mod tests {
             .iter()
             .cloned()
             .map(goldilocks_from_ark_ff)
-            .collect::<Vec<_>>()
-            .try_into()
-            .unwrap();
-
-        // Run our implementation.
-        let mut output = input;
-        poseidon2.permute_mut(&mut output);
-
-        assert_eq!(output, expected);
-    }
-
-    #[test]
-    fn test_poseidon2_babybear_width_16() {
-        const WIDTH: usize = 16;
-        const D: u64 = 7;
-        const ROUNDS_F: usize = 8;
-        const ROUNDS_P: usize = 13;
-
-        type F = BabyBear;
-
-        let mut rng = rand::thread_rng();
-
-        // Poiseidon2 reference implementation from zkhash repo.
-        let poseidon2_ref = Poseidon2Ref::new(&POSEIDON2_BABYBEAR_16_PARAMS);
-
-        // Copy over round constants from zkhash.
-        let round_constants: Vec<[F; WIDTH]> = RC16
-            .iter()
-            .map(|vec| {
-                vec.iter()
-                    .cloned()
-                    .map(babybear_from_ark_ff)
-                    .collect::<Vec<_>>()
-                    .try_into()
-                    .unwrap()
-            })
-            .collect();
-
-        // Our Poseidon2 implementation.
-        let poseidon2: Poseidon2<BabyBear, DiffusionMatrixBabybear, WIDTH, D> =
-            Poseidon2::new(ROUNDS_F, ROUNDS_P, round_constants, DiffusionMatrixBabybear);
-
-        // Generate random input and convert to both BabyBear field formats.
-        let input_u32 = rng.gen::<[u32; WIDTH]>();
-        let input_ref = input_u32
-            .iter()
-            .cloned()
-            .map(FpBabyBear::from)
-            .collect::<Vec<_>>();
-        let input = input_u32.map(F::from_wrapped_u32);
-
-        // Check that the conversion is correct.
-        assert!(input_ref
-            .iter()
-            .zip(input.iter())
-            .all(|(a, b)| babybear_from_ark_ff(*a) == *b));
-
-        // Run reference implementation.
-        let output_ref = poseidon2_ref.permutation(&input_ref);
-        let expected: [F; WIDTH] = output_ref
-            .iter()
-            .cloned()
-            .map(babybear_from_ark_ff)
             .collect::<Vec<_>>()
             .try_into()
             .unwrap();
